src: Renaming class to spell out them (Adl, Nhl, Npl, Npt, Npte)
Refs: #1534

+++ tests/test-conf-parameter.cpp
@@ -77,7 +77,7 @@ BOOST_AUTO_TEST_CASE(ConfParameterSettersAndGetters)
 
   BOOST_CHECK_EQUAL(cp1.getRootKeyPrefix(), "adminRootKey");
 
-  BOOST_CHECK_EQUAL(cp1.getInterestRetryNumber(), 2);
+  BOOST_CHECK_EQUAL(cp1.getInterestRetryNumber(), (uint32_t)2);
 
   BOOST_CHECK_EQUAL(cp1.getInterestResendTime(), 1000);
